fireban/requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-markdownx==3.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
importlib-metadata==2.0.0
Markdown==3.3
numpy==1.19.4
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.3.0
